// src/app/page/master/pegawai-struktur-gaji-by-makape/pegawai-struktur-gaji-by-makape.interface.ts
import { EventEmitter } from '@angular/core';

export interface  PegawaiStrukturGajiByMakape{
    kode?: any;
    kdGolonganPegawai?: number;
    hargaSatuan?: string
    kdOperatorRate?: string;
    kdKomponenHarga?: string;
    status1?: string;
    noRec?: string;
    label?: any;
    value?: any;
    namaStatus?: string;
    id?: any;
}